Allow hotkeys from view3d to control state of React wrapper components
The state of the camera (whether we're showing the scene through a
persp camera or an ortrographic top camera) is no longer strictly
internal to the Viewport class: we can control this state by passing
down a "camera" prop (with a string value of either "persp" or "top")
via the ReactViewport component. In componentDidUpdate we detect the
change and propagate it to our non-declarative Viewport state.