Refactor interface between Viewport and React
ViewportState now encapsulates the inputs that are fed into the
Viewport from above in order to control various things like the current
camera, which node is selected, and triggering of actions like frame
and undo.

ViewportEvents now encapsulates the state changes and other events that
are propagated back up from the Viewport, such as whether the undo stack
is populated, when the viewport is registered, and when nodes are added,
selected, or moved by dragging.